backend/requirements.txt
Flask
Flask-Cors
Flask-pymongo
python-dotenv
pytest
pytest-mock